# Structs and strings marshaling performance optimization