--- a/src/Confluent.Kafka/Impl/SafeKafkaHandle.cs
+++ b/src/Confluent.Kafka/Impl/SafeKafkaHandle.cs
@@ -406,7 +406,7 @@ namespace Confluent.Kafka.Impl
                 return false;
             }
 
-            var msg = Util.Marshal.PtrToStructure<rd_kafka_message>(msgPtr);
+            var msg = Util.Marshal.PtrToStructureUnsafe<rd_kafka_message>(msgPtr);
 
             byte[] val = null;
             if (msg.val != IntPtr.Zero)