--- a/src/Confluent.Kafka/Internal/Util.cs
+++ b/src/Confluent.Kafka/Internal/Util.cs
@@ -29,19 +29,22 @@ namespace Confluent.Kafka.Internal
             /// <summary>
             ///     Interpret a zero terminated c string as UTF-8.
             /// </summary>
-            public static string PtrToStringUTF8(IntPtr strPtr)
+            public unsafe static string PtrToStringUTF8(IntPtr strPtr)
             {
                 // TODO: Is there a built in / vectorized / better way to implement this?
                 var length = 0;
-                unsafe
-                {
-                    byte* pTraverse = (byte*)strPtr;
-                    while (*pTraverse != 0) { pTraverse += 1; }
-                    length = (int)(pTraverse - (byte*)strPtr);
-                }
+                
+                byte* pTraverse = (byte*)strPtr;
+                while (*pTraverse != 0) { pTraverse += 1; }
+                length = (int)(pTraverse - (byte*)strPtr);
+#if NET45
                 var strBuffer = new byte[length];
-                System.Runtime.InteropServices.Marshal.Copy(strPtr, strBuffer, 0, length);
+                System.Runtime.InteropServices.Marshal.Copy(strPtr, strBuffer, 0, strBuffer.Length);
                 return Encoding.UTF8.GetString(strBuffer);
+#else
+                // avoid unnecessary data copying on NET45+
+                return Encoding.UTF8.GetString((byte*)strPtr.ToPointer(), length);
+#endif
             }
 
             public static unsafe T PtrToStructure<T>(IntPtr ptr)