--- a/src/Confluent.Kafka/Producer.cs
+++ b/src/Confluent.Kafka/Producer.cs
@@ -129,7 +129,7 @@ namespace Confluent.Kafka
         /// </remarks>
         private static void DeliveryReportCallbackImpl(IntPtr rk, IntPtr rkmessage, IntPtr opaque)
         {
-            var msg = Util.Marshal.PtrToStructure<rd_kafka_message>(rkmessage);
+            var msg = Util.Marshal.PtrToStructureUnsafe<rd_kafka_message>(rkmessage);
 
             // the msg._private property has dual purpose. Here, it is an opaque pointer set
             // by Topic.Produce to be an IDeliveryHandler. When Consuming, it's for internal