Below are the results of measurements for  10 iterations per 10 mln. messages.
All values are in `msg/ms`.
`test\Confluent.Kafka.Benchmark` project was used for benchmarking.

  | average | median | std. dev
-- | -- | -- | --
produce (original) | 773,6 | 780 | 21,25653
produce (patched) | 796,8 | 805 | 20,33126
consume (original) | 548,9 | 552 | 9,658675
consume (patched) | 1722,6 | 1723 | 4,127953
poll (original) | 428,1 | 421 | 42,88461
poll (patched) | 1668,9 | 1675,5 | 28,72786

Here is a quick recap of the data above based on median values:

  | original (msg/ms) | patched (msg/ms) |  
-- | -- | -- | --
produce | 780 (+-21) | 805 (+-20) | ? might be 1.03x faster
poll | 421 (+-43) | 1675 (+-29) | ^ definitely 3.98x faster
consume | 552 (+-10) | 1723 (+-4) | ^ definitely 3.12x faster

Tests were performed on the following system:
OS: `Win10 x64 (1709)`
CPU: `Intel Core i7 6700HQ`
Kafka: `2.12-1.0.0`
Zookeeper: `3.4.11`
librdkafka: `0.11.3`
.NET Core: `2.0.3`
Java: `8`
JDK: `1.8.0_141`