@mhowlett 2018-01-17
thanks @szKarlen - everything looks good now. I want to do some additional testing (which I'm not expecting to highlight any problems), however whilst doing that I saw the integration tests are failing (unrelated to this PR). I'll merge this after getting to the bottom of that.

Anyway, awesome perf. improvement, looking forward to merging.