@mhowlett 2018-01-18
I'm getting this intermittently when running the integration tests on this branch (rebased to master). Not necessarily related to this PR but I haven't seen the issue on master. Deferring merging 'till we get to the bottom of it.

The active test run was aborted. Reason: Unhandled Exception: Xunit.Sdk.EqualException: Assert.Equal() Failure
Expected: Byte[] [1, 2, 3, 4]
Actual:   Byte[] [2, 3]
   at Xunit.Assert.Equal[T](T expected, T actual, IEqualityComparer`1 comparer) in C:\Dev\xunit\xunit\src\xunit.assert\Asserts\EqualityAsserts.cs:line 40
   at Confluent.Kafka.IntegrationTests.Tests.DeliveryHandler_P.HandleDeliveryReport(Message dr) in /git/confluent-kafka-dotnet/test/Confluent.Kafka.IntegrationTests/Tests/Producer_ProduceAsync_DeliveryHandler.cs:line 57
   at Confluent.Kafka.Producer.DeliveryReportCallbackImpl(IntPtr rk, IntPtr rkmessage, IntPtr opaque) in /git/confluent-kafka-dotnet/src/Confluent.Kafka/Producer.cs:line 155
   at Confluent.Kafka.Impl.NativeMethods.NativeMethods.rd_kafka_flush(IntPtr rk, IntPtr timeout_ms)
   at Confluent.Kafka.Impl.LibRdKafka.flush(IntPtr rk, IntPtr timeout_ms) in /git/confluent-kafka-dotnet/src/Confluent.Kafka/Impl/LibRdKafka.cs:line 665
   at Confluent.Kafka.Impl.SafeKafkaHandle.Flush(Int32 millisecondsTimeout) in /git/confluent-kafka-dotnet/src/Confluent.Kafka/Impl/SafeKafkaHandle.cs:line 136
   at Confluent.Kafka.Producer.Flush(TimeSpan timeout) in /git/confluent-kafka-dotnet/src/Confluent.Kafka/Producer.cs:line 654
   at Confluent.Kafka.IntegrationTests.Tests.Producer_ProduceAsync_DeliveryHandler(String bootstrapServers, String singlePartitionTopic, String partitionedTopic) in /git/confluent-kafka-dotnet/test/Confluent.Kafka.IntegrationTests/Tests/Producer_ProduceAsync_DeliveryHandler.cs:line 120